refactor(server): extract GameBoard from game.js into entities/gameBoard.js
game.js was 2,639 lines (three classes). GameBoard (the ~1,780-line
per-tick board/tile/ability/projectile driver) has no back-reference to
Room/Game, so it moves cleanly to server/entities/gameBoard.js, keeping
the entities dependency direction strictly downward. game.js drops to
851 lines (Room + Game only) and still exports only getRoom().

GATE_DEPTH (used only by GameBoard) moves with it. No behavior change —
smoke-test boots the real modules and ticks all 52 maps through
waiting→racing→collapsing with no crash. ARCHITECTURE.md updated.
