+++ ARCHITECTURE.md
@@ -37,7 +37,7 @@ demand → each `Room` owns a `Game` + `GameBoard` + `World` + `Engine`.
 
 | File | Lines | Purpose / contents |
 |---|---|---|
-| `game.js` | ~2.3k | Core orchestrators only: **`Room`** (per-room registry of player/projectile/aimer/hazard lists), **`Game`** (the state machine over `c.stateMap`: waiting→lobby→overview→gated→racing→collapsing→gameOver; round/brutal/scoring rules; bot fill), **`GameBoard`** (per-tick collision/ability/map driving, brutal-round config, map lifecycle). Music & achievement methods are mixed into `Game.prototype` from `music.js`/`achievements.js`. Exports only `getRoom(sig,size)`. |
+| `game.js` | ~850 | Core orchestrators only: **`Room`** (per-room registry of player/projectile/aimer/hazard lists) and **`Game`** (the state machine over `c.stateMap`: waiting→lobby→overview→gated→racing→collapsing→gameOver; round/brutal/scoring rules; bot fill). Music & achievement methods are mixed into `Game.prototype` from `music.js`/`achievements.js`. **`GameBoard`** lives in `entities/gameBoard.js` (`Room` constructs it). Exports only `getRoom(sig,size)`. |
 | `engine.js` | ~730 | Physics & collision. Per-tick updates hazards/projectiles/players; `QuadTree` broadphase against active map tiles. Helpers used from game.js/entities: `preventEscape`, `checkCollideCells`, `punchPlayer`, `puckPlayer`, `explosion`, `bounceOffBoundry`, … |
 | `utils.js` | ~500 | Central config loader (`loadConfig()` returns cached `config.json`, `PORT` override), `dt` clock, math/RNG helpers; scans `client/{maps,assets}` at boot to build the `contentDelivery` manifests; `loadMaps()` reconstructs sites-only maps to full geometry (`mapFormat`); `submitPullRequest()` (octokit) reduces an editor-submitted map to sites-only and opens a GitHub PR; shared `validateMap()`. |
 | `messenger.js` | ~210 | Socket.IO wrapper. Owns mailbox (client id→socket) and room-mailbox (id→room sig). `checkForMail()` registers every per-client handler (`enterGame`, `joinARoom`, `submitNewMap`, input events). The single place socket events are wired. |
@@ -61,6 +61,7 @@ direction is strictly downward (no cycles): `game.js` → entities → `shapes.j
 |---|---|---|
 | `shapes.js` | ~185 | Geometry primitives: `Shape` → `Rect`, `Circle`, `Gate`. Collision math (`testRect`/`testCircle`/`getExtents`). Everything else extends these. |
 | `world.js` | ~70 | `World` (extends `Rect`): bounds + player creation/spawn (`createNewPlayer`). |
+| `gameBoard.js` | ~1.8k | **`GameBoard`** — per-room map/tile/ability/projectile board state and the bulk of per-tick board logic (collision/ability driving, brutal-round config, map lifecycle). Extracted from `game.js`; no back-reference to `Room`/`Game`, so the dependency stays downward. `Room` constructs it as `new GameBoard(world, playerList, projectileList, aimerList, hazardList, engine, roomSig)`. |
 | `player.js` | ~740 | `Player` (extends `Circle`) and `LobbyStartButton`. The big behavioral class: movement, punch/puck, fire, sleep/AFK, infection, `handleHit` (type dispatcher → `handlePunchHit`/`handlePuckHit`/`handleMapCellHit`), ability pickup via the `ABILITY_TILE_CTORS` table + `tryAcquireAbility`, scoring/notches, kill/reset. |
 | `abilities.js` | ~165 | `Ability` base + `Blindfold`, `Swap`, `IceCannon`, `Bomb`, `SpeedBuff`, `SpeedDebuff`, `TileSwap`, `Cut`, `BombTrigger`. |
 | `projectiles.js` | ~145 | `Projectile` (extends `Circle`) + `CloudProj`, `SnowFlakeProj`, `BombProj`, `Puck`. |
